Add interactive CLI
Required a change to the library so that public api methods return a
`Vec<Entry>` to be used for the select list.

`seb` library is now less opionated on what the pub functions should be
doing and when, this was required to factor in the user selection from
the `Vec<Entry>` returned from an API.

Removed the `isbn` and `doi` module as the other changes made these
redundant.