Change the select list to use Entry title
The user select list of found entries now uses just the title of the
entry which makes the list easier to read.

Currently the `Entry` type doesn't enforce that title is a valid value
in the `Entry`, however #36 should add this as it is a valid value in
nearly all bibliograpic entries.