Add feature to remove entry by cite key
 # Lib

This adds the ability to remove an entry from the `Biblio` type and
introduces the `dirty` flag for the ability to check whether the
bibliography has changed and needs updating.

 # CLI
This adds the `rm` subcommand to the CLI which can be used to remove an
entry by the cite key.