Add `ResolverEntry` in-place manip of `Resolver`
Using the `ResolverEntry` makes resolving and setting the required
fields more efficient as the user no longer needs to clone the result of
the `required_fields` method to view and set the required fields.

Users can now user the `next_required_entry` method in a while let loop
in order to view the field name and provide a value which doesn't
require cloning.