Fix `Other` entry type to include custom kind text
The `Other` entry type now includes a `String` containing the type of
the custom entry type so that it can be used for bibliography formats
that have types not included in seb.

This change was also make it easier to implement #101 to include other
types then what is considered in seb.