Fix Google API response to be resolvable
The core of the code now returns a `Resolver` instead of an
`Entry::Book`. This change allows for some possible errors to be ignored
and for the field to be let unset which then can be resolved at runtime
with the user input.

Add more unit tests for the book API.