Remove concrete `Entry` for `dyn EntryExt`
The `Entry` enum before made the design to fragile and making breaking
changes were too easy, it was also impossible to extend the entry types
because of the `Entry` enum and the `EntryKind` enum.

The `Entry` type has been removed in favour of using a `Box<dyn
EntryExt>` which relies on dynamic dispatch of the trait object.

`EntryKind` type has been removed and string kind is used instead so
that the entry types can be extended too.