Fix BibTeX date to date parts
Biblatex accepts `date` field as valid and is only corrected when the
`to_bibtex_string` method is called, which doesn't help seb.

Thus when transforming the biblatex model to seb we need to normalize
a date `field` to the date parts in order to fulfill the required
fields for the resolvers.