Fix `crossref` support in BibTeX Format
biblatex lazily resolves `crossref` fields and so when converting to seb
models the raw `crossref` field would be used instead of resolving this
field to include the other data.

`biblatex` also transforms `title` from a `book` entry into `booktitle`
for an `inbook` crossref even though the `booktitle` is not correct for
`inbook` entries.