Add `file` lib feature (#143)
* Add `FormatFile` to `seb-lib` under `file` feature

Adds an improved version of `FormatFile` from `seb-cli` to the `seb-lib`
under the `file` feature flag so that it can be conditionally used.

Adds a new constructor for `Error` to wrap an existing error while still
adding a context message.

* Fix CI to test all features of `seb-lib`

`seb-lib` now contains non default features, namely the `file` feature,
which is not tested when using `cargo test --all` the `--all-features`
flag must be used.

* Replace file module code with `seb-lib` file feat.

`seb-lib` now includes `FormatFile`, which was extracted from the code
removed here, under the `file` feature. The code from `seb-lib` replaces
the existing code here.