Improve too many files found error display
This change will include all the files found whereas previously we just
listed the first and second found ignoring if we found more after that.

The error handling of possible paths are better handled now with the new
GlobIter type which wraps GlobErrors into crate errors so we can iterate
easier and just try the result with `?`.