Add new `derive` command
The `derive` command can be used to create a new entry type based on an
existing entry found in the bibliography file.
The deriving from an entry creates a new resolver of the new entry type
and sets the fields from the existing entry for that resolver then the
outstanding required fields must be resolved by the user.